rename util to localStorage
